Bugfix: test if Server::getApplications returned an array
git-svn-id: file:///home/code/svn/ovd/trunk@2435 2b2718be-6ee5-4ec5-ad1a-3fa14f9c7ca3

+++ SessionManager/web/classes/abstract/Abstract_Server.class.php
@@ -291,10 +291,12 @@ class Abstract_Server {
 			$applicationDB = new $mod_app_name();
 			
 			// remove the orphan applications
-			foreach ($apps as $an_application) {
-				if ($an_application->isOrphan()) {
-					Logger::debug('main', "Abstract_Server::delete $an_application is orphan");
-					$applicationDB->remove($an_application);
+			if (is_array($apps)) {
+				foreach ($apps as $an_application) {
+					if ($an_application->isOrphan()) {
+						Logger::debug('main', "Abstract_Server::delete $an_application is orphan");
+						$applicationDB->remove($an_application);
+					}
 				}
 			}
 		}
